Move getDomainNameservers function
