+++ src/providers/sh/util/domains/get-domain-nameservers.js
